backend/requirements.txt
Django~=5.0.3
djangorestframework~=3.15.0
django-filter
markdown
django-extensions
django-rest-swagger
drf-yasg
marko~=2.0.3